feat: Add `OnlyIf` for thrown exceptions (#25)
I often have the issue that I want to assert in a test, that an
exception was only thrown when a certain condition applies.
Currently I have to introduce an if/else clause, e.g.
```csharp
if (expectMatch)
{
    await Expect.That(sut).ThrowsNothing();
}
else
{
    await Expect.That(sut).ThrowsException().Which.HasMessage(expectedExpression);
}
```

This PR adds an `OnlyIf` method to the `DelegateExpectationResult`
class, so that this can be simplified to
```csharp
await Expect.That(sut).ThrowsException().OnlyIf(!expectMatch).Which.HasMessage(expectedExpression);
```

The `OnlyIf` will check the predicate, and if it is `true` it will
ensure, that the exception was thrown as specified. If it is `false`, it
will instead ensure, that no exception was thrown.